pyproject.toml
[project]
name = "ae-discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py[speed]>=2.5.2",
    "pydantic-settings>=2.9.1",
    "pydantic>=2.11.5",
]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]

[tool.ruff.lint]
select = [
    "A", # shadowing of Python builtins
    "B",
    "C4", # use comprehensions when possible
    "E", # pycodestyle (https://docs.astral.sh/ruff/rules/#pycodestyle-e-w)
    "ERA", # commented code
    "F", # pyflakes (https://docs.astral.sh/ruff/rules/#pyflakes-f)
    "FBT", # boolean trap
    "FLY", # f-string instead of str.join
    "FURB", # https://docs.astral.sh/ruff/rules/#refurb-furb
    "I", # isort
    "PERF", # performance
    "PLW", # pylint warnings (https://docs.astral.sh/ruff/rules/#pylint-pl)
    "RUF", # Ruff specific rules
    "SIM", # simplify (https://docs.astral.sh/ruff/rules/#flake8-simplify-sim)
    "T100", # breakpoint()
    "T2", # print statements
    "TCH", # type-checking block
    "UP", # forbid outdated syntax
]
